Add string length tests for supplementary Unicode code points
That is, code points outside the Basic Multilingual Plane (U+0000 to U+FFFF).

This will test that validators correctly account for the number of _code points_
inside strings, as required by RFC 7159, section 1:

   A string is a sequence of zero or more Unicode characters [UNICODE].
   Note that this citation references the latest version of Unicode
   rather than a specific release.  It is not expected that future
   changes in the UNICODE specification will impact the syntax of JSON.

In order to keep it "easy" for storage, the UTF-16 escape sequence (as defined
by RFC 7159, section 7) is used to represent these characters. The chosen code
point is U+1F4A9 (yes, Unicode defines that):

http://www.fileformat.info/info/unicode/char/1F4A9/index.htm