Add test case for protocol-relative uri validation
<https://tools.ietf.org/html/draft-fge-json-schema-
validation-00#section-7.3.6> states that a valid URI is whatever is
defined by <https://tools.ietf.org/html/rfc3986>. According to section
4.2 (<https://tools.ietf.org/html/rfc3986#section-4.2>), these are valid
but "rarely used" (no longer true).

At least a PHP implementation didn't support this
(<https://github.com/justinrainbow/json-schema/issues/144>), and it is a
relative edge case so it seems useful to have a test for it.