Added test cases from Foundations of JSON Schema research paper
In the 25th internationl world wide web conference there was a paper
 named "Foundations of JSON Schema", presented by:

* Felipe Pezoa (fipezoa@uc.cl)
* Juan L. Reutter (jreutter@ing.puc.cl)
* Fernando Suarez (fsuarez1@ing.puc.cl)
* Martín Ugarte (martin.ugarte@ulb.ac.be)
* Domagoj Vrgo (dvrgoc@ing.puc.cl)

See:

http://www2016.net/proceedings/proceedings/p263.pdf

This paper compared a number of json schema implementations, using a set
of 5 tests which highlighted inconsistencies between the
implementations. I have reproduced those tests here. All credit should
go to the authors of the paper for highlighting these inconsistencies
and publishing the details of them online.