Add name.json to the CLI tool's remotes.
Also fix the test that was trying to ensure that this wouldn't happen
(it was only checking that no new ones were added in-line that didn't
live on the filesystem, so now check in both directions).

Right now, remotes are maintained in both of these places, which is a
bit non-ideal, but that's the status quo. The right actual fix here is
probably to remove the in-memory schemas and serve off the filesystem
for the `remotes` command (and use shutil.copytree for dumping).