Add tests for "date" and "time" formats
This does not include any of the other RFC 3339 formats that
are reserved but not included in the main set of formats by
the validation specification.  For the same reason, it also
does not include the reserved "full-date" and "full-time" aliases.

There is no required behavior for the reserved formats.  We
may wish to add tests for them in the future, or make an
even-more-optional section of some sort.  But this should be
sufficient for the initial draft-07 test suite.