Partial IRI and IRI-reference test suite
These are just slightly changed versions of most of the URI and
URI-reference test suites.  I omitted a number of URI test cases,
particularly of different sorts of schemes.  It's not entirely
clear to me what the scope of the test suite should be, and I
just want to get something in place that hopefully someone with a
deeper understanding of IRI-related challenges can enhance.