Enforce ECMA 262 regular expression definition of $
The regular expression specification in different languages varies.

The json schema draft 3 states:

"Regular expressions SHOULD follow the regular expression specification from ECMA 262/Perl 5"
https://tools.ietf.org/html/draft-zyp-json-schema-03

In that regular expression definition:

If multiline is FALSE, $ must match the end of the string.
If multiline is TRUE, $ may match \n.

https://www.ecma-international.org/ecma-262/5.1/#sec-15.10.2.6

We should test this since Python for instance will match \n if it is the
final character even if multiline is FALSE (with \Z acting as end of
string only).