Incorporate node package and associated configuration into master branch
Previously the npm package was maintained in the "node" branch, but has
become quite outdated (that last commit was on 4th Aug 2015).
In its current state I can't see a reason why the npm specific code
can't be incorporated as part of the master branch.

This change incorporates the functionality of the node branch and adds
a linter in the form of the "standard" package.

Signed-off-by: Lucian <lucian.buzzo@gmail.com>