Make ajv's tests not fail the build.
This will shortly move to a separate job entirely, since
right now only ajv's tests are running here, and they should
not be.

But we'll likely add a way for implementations to opt-in
to having their tests run during new test additions as part
of json-schema-org/json-schema-org.github.io#6, but not fail
the build when their own tests fail.