Fix ECMA 262 regex whitespace tests.
Looks like there was a bug at the time of introduction due to
https://github.com/json-schema-org/JSON-Schema-Test-Suite/issues/285

Per ECMA 262 specification:
https://www.ecma-international.org/ecma-262/10.0/index.html#sec-characterclassescape

`\S` refers to `\s`:

> # Section 21.2.2.12:

> The production `CharacterClassEscape::S` evaluates as follows:
>  1. Return the set of all characters not included in the set returned
>     by `CharacterClassEscape::s`.

`\s` refers to `WhiteSpace`:

> The production `CharacterClassEscape::s` evaluates as follows:
>  1. Return the set of characters containing the characters that are on
>     the right-hand side of the WhiteSpace or LineTerminator
>     productions.

`WhiteSpace` includes NBSP (**U00A0**):
https://www.ecma-international.org/ecma-262/10.0/index.html#prod-WhiteSpace

That is the exact opposite of what the current test was checking for.

Also, `WhiteSpace` includes any other Unicode "Space_Separator" code
points.

That's not a new addition, same was true in version 6.0:
* https://www.ecma-international.org/ecma-262/6.0/index.html#sec-characterclassescape
* https://www.ecma-international.org/ecma-262/6.0/index.html#sec-white-space