Extend email format tests
Before this, a simple /@/ check passed this format test.

This now extends the test a bit more, to check that local part is
checked at least for something.

Symbol check + dot position check is a good indication here, because:
1) It's common and doesn't need implementing complex quotes logic
2) Dot is the only symbol which validity depends on the position in the
   local part

This should trigger most of the naive implementations (e.g. /@/) and is
still not hard to implement as it doesn't include quoted strings support.

JSON Schema uses RFC 5322.

Refs: https://tools.ietf.org/html/rfc5322#page-17
Refs: https://tools.ietf.org/html/rfc5322#page-12