Check for multipleOf overflow
The spec says that instances are valid if "dividing by this value results in an integer".

This allows integers to be invalid for `multipleOf: 0.5` if float division overflows to infinity (a non-integer); alternatively implementations may choose to implement logic which defines all integers as multiples of 0.5.  Either way, however, implementations must not raise an error due to the overflow of a legal value against a legal schema.