Add tests for $dynamicRef/$dynamicAnchor
What was there originally was copied and adapted from the 2019-09 test.
While those tests are valid, I thought it would be easier for people to
follow this complicated concept by having just two examples that are
changed slightly to cover all the normal cases and edges cases related
to these keywords. That ended up creating some duplication between some
of the old tests with some of the new tests. I favored the new tests
over the old ones when there was conflict.