Add another test case for unevaluatedItems-contains interaction
We use contains and unevaluatedItems to define a dependency relationship over the array items. The array may contain any sequence of a's, b's and c's, but b's may only be present if there is at least one a, and c's may only be present if there is at least one b.