Revert "by default, "format" only annotates, not validates"
This reverts commit 043dc63692a8c9533772dd771cdb32742ca503d2
(and also removes the tests from draft-future, which were copied from
draft2020-12).

According to the spec, "$vocabulary": { "...format-assertion": false }
does not mean "do not assert formats", but rather "if you don't support
format-assertion, you can still load the schema. but if you do support it,
feel free to assert that behaviour if you wish".

ref. https://json-schema.org/draft/2020-12/json-schema-core.html#rfc.section.8.1.2

As such, the data in the removed tests could result in either a true OR a
false result. If we wish to explicitly test "formats are only annotations, and
NOT assertions" behaviour, we will have to define a custom metaschema and
reference that with the $schema keyword, where the metaschema contains:

    "$vocabulary": {
        ...
        "https://json-schema.org/draft/2020-12/vocab/format-annotation": true,
    }

and does NOT mention the format-assertion vocabulary whatsoever.

This would be a reasonable test of the $vocabulary keyword, too.