Remove the broken and obsolete JavaScript-based test integration
Looks like this repo has had for a very long time a Mocha and AJV based
test runner implemented on `index.js`.

This test is completely broken right now and relies on a quite old AJV
version. Furthermore, I think the JSON Schema test suite should remain
implementation-agnostic and not i.e. bundle with AJV or other
implementations itself.

If this is the case, in the interest of keeping the repo on a clean
state, would it make sense to just delete this dead code and bring it
back in the future if really needed?

Signed-off-by: Juan Cruz Viotti <jv@jviotti.com>