Revert PR#376.
The final form of these files as merged into main is not at all what was
intended in the original PR, and the tests are not correct: originally, they
were tests taken from later drafts of keywords that were introduced in these
later drafts, but with all the results changed to 'true' in order to
demonstrate that the keywords were NOT recognized - that they were entirely
ignored. The tests were moved to optional/, apparently with the intent to
express that an implementation might wish to recognize those keywords anyway
(which seems odd, but could be worth a discussion to see if this would be
useful), but the results were never changed from being all 'true' back to the
actual true-or-false results as in the specification version where they were
introduced. Consequently they are not valid tests and should be removed.