Update the test schema to Draft 2020, and fix a bug in it.
`minItems` was placed inside `items` in the root schema, meaning it was
not checking that each file contained at least one test case. Now it
does, and does so using draft 2020.

Also tidied a few descriptions, removed two "dead properties"
(id, output) which we added in prep for enriching the suite but haven't
yet added concrete examples of. We can re-add when we do add some.