Forward-port and reword the new unevaluatedItems test.
Makes it fit a bit better with its neighbors by using similar language
in the descriptions.

This *precise* case seems uncovered by said neighbors
because none precisely have all of:

    * nested unevaluatedItems
    * no prefixItems
    * cases where unevaluatedItems actually gets used for a scenario

It's possible that in aggregate all the cases that are here already
would catch any bugs this scenario would catch (i.e. it's possible this
is redundant in aggregate) but it's safe enough to add just in case
obviously.