Also check case description length in sanity checks.
The reason we do this is both to force conciseness but also because
runners of the suite likely turn one or more of the combination of
case description and test description into an identifier within the
implementer's test framework. If descriptions are really long, it makes
things like copy pasting test names more annoying or difficult.

150 is quite long -- we use 70 for test descriptions, so in total this
means potentially 220 characters for the combined length. I'd like to
bring this down a bit but this is the number that passes all existing
descriptions at least.