Add a badge for a Zenodo DOI.
This allows for easier citations for researchers using the suite.

This DOI will change each time a release is published. How often that
happens is for the moment still ad hoc (i.e. we continually add tests,
and will at some period decide to batch them into a release). In the
future we may decide to automatically publish releases on some cadence.