Emit fully absolute URIs from `jsonschema_suite remotes`.
This was I believe just an oversight on my part (or I never noticed
it before) -- but there's no reason a user of this should need to
know the base URI is http://localhost:1234/, the whole point of the
command is to just give you the URIs + schemas to configure.

I don't know how many people use this command rather than reading
directly off the remotes/ directory, but I do, so it's at least 1
person.