Optional tests for no $schema (2019-09+)
minLength is identical across all drafts, and highly unlikely
to ever change, so regardless of how an implementation chooses
to process this schema, if it processes it at all the behavior
should be consistent.

Also, TIL that "type" has *not* always been the same - in the first
few drafts it was always an array of types *or schemas*!  While
we don't even have test suites for those drafts, it seemed better
to use a keyword that has never changed.