Add in a contributors document outlining the repo's process.                                                                                                                julian@Airm
This document is undoubtedly incomplete and can be iterated on.

It represents, to the best of my knowledge, the *existing* process
I have followed myself since the repository's inception, with only
minor tweaks (back then lots of GitHub UI features didn't even
exist!).

Of course over the years it may not have been followed by everyone,
nor even possibly by myself, but putting it here should hopefully
make it clearer to everyone what the baseline is, in case we wish
to make changes.

There are some differences here from processes followed in other JSON
Schema org repos -- whether this is an issue or not is of course
constructively debatable.