Remove unneeded remotes
I had copied these blindly from the top-level remotes directory,
but they are never referenced in this location.  The 2019-09,
2020-12, and next tests reference the loctionIndependentIdentifier.json
files in these subdirectories.  They do not reference the ...Draft4
or ...Pre2019 files at all.

Tests for earlier drafts reference the ...Draft4 and ...Pre2019 files
in the root remotes directory only.