Ensure we don't use unknown keywords unless intended.
This adds a sanity check which ensures a test doesn't use
an unknown keyword unless it's a test specifically meant to
test behavior with unknown keywords.

The way it does so is super naively -- with a hardcoded
list of keywords from each draft (which are basically monkey
patched in), and then skipping tests which mention that
they have something to do with unknown keywords.

It's likely this list may need tweaking at some point if
someone creates such a test that doesn't match the allowlist
but overall it catches a (very old draft 3) test which is just
wrong, so it seems likely it's helpful to have.