Fix a sloppy bug in the sanity checks in CI.
This was sloppy -- and was never running :/ because it was mis-indented,
causing us to not actually check any schemas for validity.

This is part of why for loops in tests are so bad, but we are a bit
loose in this file (particularly because I've avoided adding
dependencies so it's slightly easier to run for those who aren't Python
developers). We should possibly refactor a bit anyhow so that an error
is raised if the for loops are empty.