Ensure $ids are collected within if/then/else schemas.
This is the first of probably quite a few trivial changes which
make our testing of $ref more resillient -- specifically today,
we only test very few keywords to ensure subschemas within them
are collected / searched for $ids.

if/then/else are possibly particularly dangerous, since some
tools may mistakenly believe that a schema with no if can have
then/else dropped, but this is not the case, as the then/else
may contain a subschema that will be referenceable elsewhere.

Found via discussion with Giorgio Ghelli.